syntax check

+++ src/channel.cpp
@@ -49,7 +49,7 @@ void    Channel::leftTheChannel(Client *client)
 std::string     Channel::getUsers()
 {
     std::string users;
-    for (int i = 0 ; i < _channelClients.size() ; i++)
+    for (size_t i = 0 ; i < _channelClients.size() ; i++)
 	    users.append(_channelClients[i]->getNickName() + " ");
     return users;
 }
